+++ CHANGELOG.md
@@ -7,9 +7,11 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [unreleased]
+
 ## [0.0.53] - 2021-01-03
 ### Add
-- show `--validate` file-name/ -path
+- Show `--validate` file-name/ -path
 ### Change
 - Reduce of past-tense in change log headlines
 ### Fix